reference-implementation/python/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "a2cn"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.115.0",
    "uvicorn==0.32.0",
    "httpx==0.27.0",
    "PyJWT==2.9.0",
    "cryptography==43.0.0",
    "jcs==0.2.1",
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.0",
    "pytest-asyncio==0.24.0",
    "python-multipart==0.0.12",
    "respx==0.21.1",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["a2cn*"]
